refactor: 체스 게임 실행 동작을 별도의 클래스로 분리
체스 기물의 움직임, 포인트 계산을 `ChessGame` 클래스로 로직을 이동시킴.
