+++ src/main/java/com/seong/Main.java
@@ -1,6 +1,7 @@
 package com.seong;
 
 import com.seong.chess.Board;
+import com.seong.chess.ChessGame;
 import java.util.Scanner;
 
 public class Main {
@@ -12,17 +13,18 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         Board board = new Board();
+        ChessGame chessGame = new ChessGame(board);
         while (sc.hasNext()) {
             String command = sc.nextLine();
             if (command.equals(START)) {
-                board.initialize();
+                chessGame.initialize();
             } else if (command.equals(END)) {
                 break;
             } else if (command.startsWith(MOVE)) {
                 String[] commands = command.split(" ");
                 String sourcePosition = commands[1];
                 String targetPosition = commands[2];
-                board.move(sourcePosition, targetPosition);
+                chessGame.move(sourcePosition, targetPosition);
             }
             System.out.println(board.showBoard());
         }
