+++ src/main/java/com/seong/chess/Board.java
@@ -2,7 +2,6 @@ package com.seong.chess;
 
 import static com.seong.chess.utils.StringUtils.appendNewLine;
 
-import com.seong.chess.pieces.Blank;
 import com.seong.chess.pieces.Piece;
 import com.seong.chess.pieces.Piece.Color;
 import com.seong.chess.pieces.Piece.Type;
@@ -63,31 +62,12 @@ public class Board {
         ranks.get(position.row()).move(position.col(), piece);
     }
 
-    public void move(String sourcePosition, String targetPosition) {
-        Piece piece = findPiece(sourcePosition);
-        move(sourcePosition, Blank.create());
-        move(targetPosition, piece);
-    }
-
     public Piece findPiece(String rawPosition) {
         Position position = Position.convert(rawPosition);
         return ranks.get(position.row()).get(position.col());
     }
 
-    public double calculatePoint(Color color) {
-        Double result = ranks.stream()
-                .flatMap(rank -> rank.getSameColorPieces(color).stream())
-                .filter(piece -> !piece.isEqual(Type.PAWN, color))
-                .reduce(0D, (point, piece) -> point + piece.getDefaultPoint(), Double::sum);
-
-        for (int i = 0; i < BOARD_LENGTH; i++) {
-            double pawnCount = getColumnPawnCount(color, i);
-            result += pawnCount * Type.PAWN.getDefaultPoint();
-        }
-        return result;
-    }
-
-    private double getColumnPawnCount(Color color, int row) {
+    public double getColumnPawnCount(Color color, int row) {
         double pawnCount = 0;
         for (int col = 0; col < BOARD_LENGTH; col++) {
             Piece piece = ranks.get(col).get(row);
