+++ src/test/java/com/seong/chess/ChessGameTest.java
@@ -1,0 +1,58 @@
+package com.seong.chess;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import com.seong.chess.pieces.Piece;
+import com.seong.chess.pieces.Piece.Color;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class ChessGameTest {
+
+    private Board board;
+    private ChessGame chessGame;
+
+    @BeforeEach
+    void setUp() {
+        board = new Board();
+        chessGame = new ChessGame(board);
+    }
+
+    @Test
+    @DisplayName("체스 보드의 기물은 현재 위치에서 다른 위치로 이동할 수 있다.")
+    public void move() throws Exception {
+        chessGame.initialize();
+
+        String sourcePosition = "b2";
+        String targetPosition = "b3";
+        chessGame.move(sourcePosition, targetPosition);
+        assertEquals(Piece.createBlank(Position.convert(sourcePosition)), board.findPiece(sourcePosition));
+        assertEquals(Piece.createWhitePawn(Position.convert(targetPosition)), board.findPiece(targetPosition));
+    }
+
+    @Test
+    @DisplayName("체스판 위의 기물에 따라 점수를 계산할 수 있다.")
+    public void calculatePoint() throws Exception {
+        board.initializeEmpty();
+
+        addPiece("b6", Piece.createBlackPawn());
+        addPiece("e6", Piece.createBlackQueen());
+        addPiece("b8", Piece.createBlackKing());
+        addPiece("c8", Piece.createBlackRook());
+
+        addPiece("f2", Piece.createWhitePawn());
+        addPiece("g2", Piece.createWhitePawn());
+        addPiece("e1", Piece.createWhiteRook());
+        addPiece("f1", Piece.createWhiteKing());
+
+        assertEquals(15.0, chessGame.calculatePoint(Color.BLACK), 0.01);
+        assertEquals(7.0, chessGame.calculatePoint(Color.WHITE), 0.01);
+
+        System.out.println(board.showBoard());
+    }
+
+    private void addPiece(String position, Piece piece) {
+        board.move(position, piece);
+    }
+}
